+++ kotlin/maui-kmp-ksp/src/main/kotlin/de/voize/mauikmp/ksp/processor/csharp/CSharp.kt
@@ -1,62 +1,61 @@
 package de.voize.mauikmp.ksp.processor.csharp
 
-import de.voize.mauikmp.ksp.processor.NSArrayClassName
+import de.voize.mauikmp.ksp.processor.ArrayClassName
 import de.voize.mauikmp.ksp.processor.cSharpAttributesToString
 
 object CSharp {
     sealed interface TypeName : Annotatable {
-        fun writeTo(
-            out: Appendable,
-            withAttributes: Boolean,
-        )
-    }
+        val isNullable: Boolean
 
-    data class ClassName(
-        val simpleName: String,
-        val namespace: String,
-        val isNullable: Boolean,
-        override val attributes: List<String> = emptyList(),
-    ) : TypeName {
-        override fun writeTo(
-            out: Appendable,
-            withAttributes: Boolean,
-        ) {
-            if (withAttributes && attributes.isNotEmpty()) {
+        fun writeTo(out: Appendable)
+
+        fun writeAttributesTo(out: Appendable) {
+            if (attributes.isNotEmpty()) {
                 out.append(attributes.cSharpAttributesToString())
                 out.append(" ")
             }
-            if (namespace.isNotEmpty()) {
-                out.append(namespace)
-                out.append(".")
-            }
-            out.append(simpleName)
+        }
+
+        fun writeNullableTo(out: Appendable) {
             if (isNullable) {
                 out.append("?")
             }
         }
     }
 
+    data class ClassName(
+        val simpleName: String,
+        val namespace: String,
+        override val isNullable: Boolean,
+        override val attributes: List<String> = emptyList(),
+    ) : TypeName {
+        override fun writeTo(out: Appendable) {
+            if (namespace.isNotEmpty()) {
+                out.append(namespace)
+                out.append(".")
+            }
+            out.append(simpleName)
+        }
+    }
+
     data class ParameterizedTypeName(
         val rawType: ClassName,
         val typeArguments: List<TypeName>,
+        override val isNullable: Boolean,
     ) : TypeName {
-        override fun writeTo(
-            out: Appendable,
-            withAttributes: Boolean,
-        ) {
-            if (rawType.simpleName == NSArrayClassName.simpleName) {
-                if (withAttributes && attributes.isNotEmpty()) {
-                    out.append(attributes.cSharpAttributesToString())
-                    out.append(" ")
-                }
-                typeArguments.single().writeTo(out, withAttributes = false)
+        override fun writeTo(out: Appendable) {
+            if (rawType.simpleName == ArrayClassName.simpleName && rawType.namespace == ArrayClassName.namespace) {
+                val elementType = typeArguments.single()
+                elementType.writeTo(out)
+                elementType.writeNullableTo(out)
                 out.append("[]")
             } else {
-                rawType.writeTo(out, withAttributes)
+                rawType.writeTo(out)
                 if (typeArguments.isNotEmpty()) {
                     out.append("<")
                     typeArguments.forEachIndexed { index, typeName ->
-                        typeName.writeTo(out, withAttributes = false)
+                        typeName.writeTo(out)
+                        typeName.writeNullableTo(out)
                         if (index < typeArguments.size - 1) {
                             out.append(", ")
                         }